♻️ refactor(styles): move ScreenResponse & add Theme in styles directory (#38)
theme 생성 및 screenResponse styles 폴더로 이동

#38
